+++ RELEASE_NOTES.md
@@ -1,7 +1,23 @@
 ## 0.1.0-SNAPSHOT
-User repository created
+User repository and service created
 - [x] User repository
 - [x] Login
 - [x] Logout
 - [x] JWT authentication
 - [x] User Controller
+
+## 0.2.0-SNAPSHOT
+Product category and service created
+- [x] Product category
+- [x] Product category CRUD
+- [x] Product category Controller
+
+Product repository and service created
+- [x] Product repository
+- [x] Product CRUD
+- [x] Product Controller
+
+Order, Product in Orders repositories and services created
+- [x] Product in Orders, Orders repositories
+- [x] Product in Orders, Orders CRUD
+- [x] Product in Orders, Orders Controller